# test: Speed up auth tests