@jkosh44 2023-10-03
This should ideally lower the time that `wait_for_refresh` takes to ~0 seconds (really however long a refresh takes) since we are constantly refreshing a token immediately after getting it. Additionally it shouldn't reduce test flakiness since we still have 20 seconds from the time a refresh starts to the time a token expires.

Locally I see the following speedups:
```
test_auth_expiry:
  main: 1 min 50 sec
  PR: 30 sec
test_auth_admin_superuser_revoked (formally test_auth_admin_show_is_superuser):
  main: 49 sec
  PR: 9 sec
```