@jkosh44 2023-10-03
> How about: `const REFRESH_BEFORE_SECS: u64 = EXPIRES_IN_SECS - 5;`? If we are currently waiting 20s for each refresh, we'd now be waiting 5s for each refresh, and a 15s gain per refresh would still make the tests much faster, and allow us to sit in the "not yet refreshing" state for a while.

I think that's a good compromise, just pushed an update.

> What do you think about removing all the duplication and only having one copy of these consts?

I think that would be cleaner, just pushed an update.